# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar-IndigoPay API
  version: 1.0.0
  description: |
    OpenAPI 3.0 documentation for the backend REST API used by Stellar-IndigoPay.
    The API exposes project discovery, donations, profiles, leaderboard, notifications,
    ratings, updates, jobs, stats, and admin management endpoints.
  contact:
    name: Stellar-IndigoPay Team
    url: https://github.com/Stellar-IndigoPay/Stellar-IndigoPay
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development server  
  - url: https://api.stellarindigopay.com
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      description: Project identifier.
      schema:
        type: string
...
    PublicKeyParam:
      name: publicKey
      in: path
      required: true
      description: Stellar public key.
      schema:
        type: string
        pattern: "^G[A-Z0-9]{55}$"
...
  responses:
    BadRequest:
      description: The request payload or query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimitExceeded:
      description: The request has been rate-limited. All endpoints carry rate-limit response headers (X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset) and return 429 with a Retry-After header when the limit is exceeded. See the API documentation for per-endpoint tiers.
      headers:
        Retry-After:
          description: Number of seconds to wait before retrying.
          schema:
            type: integer
        X-RateLimit-Limit:
          description: Maximum requests allowed (or bucket capacity for token-bucket endpoints).
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Requests remaining in the current window (or tokens left in the bucket).
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Epoch seconds when the limit resets (or next token becomes available).
          schema:
            type: integer
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Too many requests — Try again later.
              retryAfter:
                type: integer
                description: Epoch seconds after which a retry may succeed.
                example: 1742169602
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request payload
      required:
        - error
...
    Donation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        donorAddress:
          type: string
        amount:
          type: string
        amountXLM:
          type: string
          nullable: true
        currency:
          type: string
        message:
          type: string
          nullable: true
        transactionHash:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CreateDonationRequest:
      type: object
      required: [projectId, donorAddress, amountXLM, transactionHash]
      properties:
        projectId:
          type: string
          format: uuid
        donorAddress:
          type: string
          pattern: "^G[A-Z0-9]{55}$"
        amountXLM:
          type: number
          minimum: 0.0000001
        amount:
          type: number
        currency:
          type: string
          default: XLM
        message:
          type: string
          maxLength: 100
        transactionHash:
          type: string
          pattern: "^[a-fA-F0-9]{64}$"
...
paths:
...
  /api/donations:
    post:
      tags: [Donations]
      summary: Record a donation
      operationId: recordDonation
      description: |
        Record a new donation made to a project, including the transaction hash and amount.

        **Idempotency:** Clients can safely retry this endpoint by supplying an
        `Idempotency-Key` header with a UUID v4 value. The server stores the
        response keyed by this header and replays it on subsequent requests with
        the same key within 24 hours. If the key is missing or malformed the
        request is processed normally (no idempotency guarantee).
      security: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: A UUID v4 that uniquely identifies this donation request. Supplying the same key within 24 hours returns the original response without creating a duplicate record.
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDonationRequest"
      responses:
        "200":
          description: Donation already recorded (idempotent replay).
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Donation"
        "201":
          description: Donation recorded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Donation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
  /api/donations/project/{projectId}/messages:
    get:
      tags: [Donations]
      summary: List donation messages for a project
      operationId: listDonationMessages
      description: Retrieve donation messages left by donors for a specific project.
      parameters:
        - $ref: "#/components/parameters/ProjectIdParam"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        "200":
          description: Donation messages.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Donation"
  /api/donations/project/{projectId}:
    get:
      tags: [Donations]
      summary: List donations for a project
      operationId: listProjectDonations
      description: Retrieve a paginated list of all donations made to a specific project.
      parameters:
        - $ref: "#/components/parameters/ProjectIdParam"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Project donation history.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Donation"
                  nextCursor:
                    type: string
                    nullable: true
  /api/donations/donor/{publicKey}:
    get:
      tags: [Donations]
      summary: List donations by donor
      operationId: listDonorDonations
      description: Retrieve all donations made by a specific donor, identified by their Stellar public key.
      parameters:
        - $ref: "#/components/parameters/PublicKeyParam"
      responses:
        "200":
          description: Donor donation history.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Donation"
        "400":
          $ref: "#/components/responses/BadRequest"
...